localize match paths function
